[Refactor] Unify Snake/SnakeBeta and alias-free activation into common modules
Consolidate duplicated Snake/SnakeBeta activation and anti-aliased
Activation1d (upsample+snake+downsample) implementations from five
speech decoder models into shared common/ modules, reducing ~600 lines
of duplication while preserving NPU bf16 workarounds and Triton dispatch.

RFC: #3884

- Add Snake subclass to common/snake_activation.py (single-param variant
  for CosyVoice3, derives inv_beta from alpha, checkpoint-compatible)
- Create common/alias_free_activation.py with AliasFreeActivation1d,
  UpSample1d, DownSample1d, and Kaiser sinc filter utilities
- Include NPU bf16 workarounds (replication_pad_1d, CPU kaiser_window)
  and XPU kaiser_window fallback in shared module
- Replace local copies in qwen2_5_omni, qwen3_tts_v1, covo_audio,
  cosyvoice3 with imports from common modules
- Clean up unused imports across all modified files
